[Chore 161638906] Update admin role validation

+++ app/api/V2/views/products.py
@@ -1,10 +1,10 @@
 from flask_restful import Resource, reqparse
 from flask import request, json, jsonify, make_response
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from functools import wraps
 
-from app.api.V2.models import ProductModel
-from app.api.V2.views.users import admin_only
+from app.api.V2.models import ProductModel, UserModel
+
 
 parser = reqparse.RequestParser()
 parser.add_argument('name', required=True, help='Product name cannot be blank', type=str)
@@ -14,13 +14,14 @@ parser.add_argument('inventory', required=True, help="Define available stock", t
 parser.add_argument('minimum_stock', required=True, help="Define minimum stock", type=int)
 
 
-class Products(Resource, ProductModel):
+class Products(Resource, ProductModel, UserModel):
 
     def __init__(self):
         self.operation = ProductModel()
 
     @jwt_required
     def get(self):
+
         products = self.operation.get_all_products()
         if not products:
             return {"message": "No products yet"}
@@ -38,6 +39,9 @@ class Products(Resource, ProductModel):
         inventory = args.get('inventory')
         minimum_stock = args.get('minimum_stock')
         try:
+            user = UserModel.find_by_email(get_jwt_identity())
+            if user[4] != "admin":
+                return {"message": "You do not have authorization to access this feature"}
             product = self.operation.get_item_if_exists(name, price, inventory, minimum_stock, category)
             return product
         except Exception as my_exception:
@@ -45,13 +49,16 @@ class Products(Resource, ProductModel):
             return {'message': 'Something went wrong.'}, 500
 
 
-class SingleProduct(Resource, ProductModel):
+class SingleProduct(Resource, ProductModel, UserModel):
 
     def get(self, id):
         return ProductModel.get_each_product(self, id)
 
     @jwt_required
     def delete(self, id):
+        user = UserModel.find_by_email(get_jwt_identity())
+        if user[4] != "admin":
+            return {"message": "You do not have authorization to access this feature"}
         return ProductModel.delete_product(self, id)
 
     @jwt_required
@@ -63,4 +70,7 @@ class SingleProduct(Resource, ProductModel):
         inventory = args.get('inventory')
         minimum_stock = args.get('minimum_stock')
 
+        user = UserModel.find_by_email(get_jwt_identity())
+        if user[4] != "admin":
+            return {"message": "You do not have authorization to access this feature"}
         return ProductModel.update_product(self, id, price, inventory, minimum_stock, category)
